requirements.txt
javascript
python-dotenv
openai
anthropic
google-generativeai
huggingface_hub
pytest
